requirements.txt
tensorflow>=1.9.0
Keras>=2.2.0
numpy>=1.14.3
scikit-learn>=0.19.1
seqeval>=0.0.5
JPype1>=0.6.3
-e git+https://www.github.com/keras-team/keras-contrib.git#egg=keras-contrib